Add `relays_mut` fn
This fixes the problem when the relays were being connected, but
they wouldn't show up in the HashMap. The problem was that we're
cloning the locked hashmap of relays and then trying to insert into
that clone.

When you do that, the actual Arc<Mutex<HashMap<String, RelayData>>>
would not be changed. Therefore, any attempt to send something to those
relays that were added after the Client was created, would not be
successful as they weren't there.