Change table reading impl to differentiate between tests and not tests
Now when reading to subscriptions_table or keys_table, it will
differentiate between tests (#[cfg(test)]) and
not tests (#[cfg(not(test))]). This will allow tests to pass a `dir`
and `file` variable, and therefore we will be able to run the tests in
parallel without worrying about using the same database connection.